Refactor: move specific service logic from utils to helpers
